Исправление недочетов
Удаление лишних файлов, перенос системных данных в utils, избавление от styles, различные правки
