+++ src/components/app/app.module.css
@@ -1,37 +1,27 @@
+.app {
+    text-align: center;
+    /* margin-left: 20rem;
+    margin-right: 30rem; */
+    width: 100%;
+  }
+  
+  .app-header {
+    background-color: #282c34;
+    
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    font-size: 24px;
+    color: white;
+    text-align: center;
+    vertical-align: middle;
+  }
+  
+
 .headerItem {
     height: 50px;
 }
-
-.column {
-    align-items: center;
-    flex-direction: column;
-
-}
-
-.ingredientsList {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-}
-
-.leftColumn {
-    composes: column;
-    margin-top: 40px;
-    margin-right: 20px;
-    width: 600px;
-
-}
-
-.rightColumn {
-    composes: column;
-
-    /* overflow-y: scroll; */
-    width: 500px;
-    margin-top: 100px;
-    margin-left: 20px;
-
-}
-
 
 .main {
     display: flex;
