+++ src/components/burger-constructor/burger-constructor.module.css
@@ -17,4 +17,39 @@
   display: flex;
   justify-content: center;
   align-items: center;
+  margin-right: 2px;
+  margin-top: 16px;
+  margin-bottom: 16px;
+}
+
+.borderItems{
+  composes: constructorIngregient;
+  margin-top: 10px;
+  margin-bottom: 10px;
+  margin-left:32px;
+  margin-right: 2px;
+
+}
+
+.checkout{
+  margin: auto;
+  margin-top: 40px;
+  display: flex;
+  justify-content: right;
+}
+
+.price{
+  margin-right: 40px;
+  margin-top: 10px;
+}
+
+
+
+.rightColumn {
+  align-items: center;
+  flex-direction: column;
+  width: 600px;
+  margin-top: 100px;
+  margin-left: 20px;
+
 }
