+++ src/components/burger-ingredients/burger-ingredients.js
@@ -1,92 +1,91 @@
 import React from 'react';
 import ingredientStyle from './burger-ingredients.module.css';
 import PropTypes from 'prop-types';
+import TabHeader from '../tab-header/tab-header';
 
 import { CurrencyIcon, Counter } from '@ya.praktikum/react-developer-burger-ui-components';
+import { ingredientPropTypes } from '../../utils/types';
 
 function BurgerIngredients({ data }) {
 
     return (
-        <section className={ingredientStyle.mainIngredientList}>
-            <div className={ingredientStyle.ingredientList}>
-                <p className="text text_type_main-medium" style={{ textAlign: 'left' }}>Булки</p>
-                {
-                    data.filter((item) => {
-                        if(item.type==='bun')
+        <div className={ingredientStyle.leftColumn}>
+            <h1 className={`${ingredientStyle.title} text text_type_main-large`} >Соберите бургер</h1>
+            <TabHeader />
+            <div className={ingredientStyle.ingredientsList}>
+                <section className={ingredientStyle.mainIngredientList}>
+                    <div >
+                        <p className={`${ingredientStyle.title} text text_type_main-medium`} >Булки</p>
                         {
-                            return item;
+                            data.filter((item) => {
+                                if (item.type === 'bun') {
+                                    return item;
+                                }
+                            })
+                                .map((item) => (
+                                    <div className={ingredientStyle.ingredient} key={item._id}>
+                                        <Counter count={1} size="default" />
+                                        <img src={item.image} className={ingredientStyle.priceBlock}></img>
+                                        <p className="text text_type_main-default">{item.price}
+                                            <CurrencyIcon type="primary" />
+                                        </p>
+                                        <p className="text text_type_main-default">{item.name}</p>
+                                    </div>
+                                )
+                                )
                         }
-                    })
-                    .map((item, key) => (
-                        <div className={ingredientStyle.ingredient} key={key}>
-                            <Counter count={1} size="default" />
-                            <img src={item.image} className={ingredientStyle.priceBlock}></img>
-                            <p className="text text_type_main-default">{item.price} 
-                                <CurrencyIcon type="primary" />
-                            </p>
-                            <p className="text text_type_main-default">{item.name}</p>
-                        </div>
-                        )
-                    )
-                }
-            </div>
-            <div className={ingredientStyle.ingredientList}>
-                <p className="text text_type_main-medium" style={{ textAlign: 'left' }}>Соусы</p>
-                {
-                    data.filter((item) => {
-                        if(item.type==='sauce')
+                    </div>
+                    <div className={ingredientStyle.ingredientList}>
+                        <p className={`${ingredientStyle.title} text text_type_main-medium`}>Соусы</p>
                         {
-                            return item;
+                            data.filter((item) => {
+                                if (item.type === 'sauce') {
+                                    return item;
+                                }
+                            })
+                                .map((item) => (
+                                    <div className={ingredientStyle.ingredient} key={item._id}>
+                                        <Counter count={1} size="default" />
+                                        <img src={item.image} className={ingredientStyle.priceBlock}></img>
+                                        <p className="text text_type_main-default">{item.price}
+                                            <CurrencyIcon type="primary" />
+                                        </p>
+                                        <p className="text text_type_main-default">{item.name}</p>
+                                    </div>
+                                )
+                                )
                         }
-                    })
-                    .map((item, key) => (
-                        <div className={ingredientStyle.ingredient} key={key}>
-                            <Counter count={1} size="default" />
-                            <img src={item.image} className={ingredientStyle.priceBlock}></img>
-                            <p className="text text_type_main-default">{item.price} 
-                                <CurrencyIcon type="primary" />
-                            </p>
-                            <p className="text text_type_main-default">{item.name}</p>
-                        </div>
-                        )
-                    )
-                }
-            </div>
-            <div className={ingredientStyle.ingredientList}>
-                <p className="text text_type_main-medium" style={{ textAlign: 'left' }}>Начинка</p>
-                {
-                    data.filter((item) => {
-                        if(item.type==='main')
+                    </div>
+                    <div className={ingredientStyle.ingredientList}>
+                        <p className={`${ingredientStyle.title} text text_type_main-medium`}>Начинка</p>
                         {
-                            return item;
+                            data.filter((item) => {
+                                if (item.type === 'main') {
+                                    return item;
+                                }
+                            })
+                                .map((item) => (
+                                    <div className={ingredientStyle.ingredient} key={item._id}>
+                                        <Counter count={1} size="default" />
+                                        <img src={item.image} className={ingredientStyle.priceBlock}></img>
+                                        <p className="text text_type_main-default">{item.price}
+                                            <CurrencyIcon type="primary" />
+                                        </p>
+                                        <p className="text text_type_main-default">{item.name}</p>
+                                    </div>
+                                )
+                                )
                         }
-                    })
-                    .map((item, key) => (
-                        <div className={ingredientStyle.ingredient} key={key}>
-                            <Counter count={1} size="default" />
-                            <img src={item.image} className={ingredientStyle.priceBlock}></img>
-                            <p className="text text_type_main-default">{item.price} 
-                                <CurrencyIcon type="primary" />
-                            </p>
-                            <p className="text text_type_main-default">{item.name}</p>
-                        </div>
-                        )
-                    )
-                }
+                    </div>
+
+                </section>
             </div>
-            
-        </section>
+        </div>
     )
 }
 
-BurgerIngredients.propTypes ={
-    data : PropTypes.arrayOf(PropTypes.shape({
-        name : PropTypes.string.isRequired,   
-        image : PropTypes.string.isRequired,   
-        price : PropTypes.number.isRequired   
-    }).isRequired
-
-    )
-  }
+BurgerIngredients.propTypes = {
+    data: PropTypes.arrayOf(ingredientPropTypes).isRequired,
+}
 
 export default BurgerIngredients;
