+++ src/components/tab-header/tab-header.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import { Tab } from '@ya.praktikum/react-developer-burger-ui-components';
+import tabStyles from './tab-header.module.css';
 
 
 function TabHeader () {
     const [current, setCurrent] = React.useState('one')
     return (
-      <div style={{ display: 'flex' }}>
+      <div className={tabStyles.tabDiv}>
         <Tab value="one" active={current === 'one'} onClick={setCurrent}>
           Булки
         </Tab>
