+++ src/components/tab-header/tab-header.module.css
@@ -1,0 +1,3 @@
+.tabDiv{
+    display: flex;
+}
