+++ src/utils/types.js
@@ -1,0 +1,8 @@
+import PropTypes from 'prop-types';
+
+export const ingredientPropTypes = PropTypes.shape({
+    _id:PropTypes.string.isRequired,
+    name: PropTypes.string.isRequired,
+    image: PropTypes.string.isRequired,
+    price: PropTypes.number.isRequired
+})
